generate.py: Enable the detection of deprecated config element
From release to release, some configuration environment my disappear
or get renamed. Schema validation is an option to confirm that the
yaml file is valid, but the generate.py should be able to let the
integrator know which variable is deprecated, hence s/he will be
able to take action upon.