configure: restart apache after disabling puppetmaster
Before: Apache was configured, puppetdb started and Apache started.
It led to some issues during the upgrade where Apache did not restart
because it was already started and needed to be restarted for some
configurations changes reasons.

Now: Apache is configured, puppetdb started and Apache restarted to
ensure Puppet will run correctly.