download.sh: Allow provision to retrieve puppet modules from tar ball
Currently the download.sh script retrieves the puppet modules running
r10k. Unfortunately, we - as consumer - don't have full control about
what happens upstream, and if a git push -f comes in the way, the r10k
for a given release won't be working anymore. Since we already tar those
modules, we offer the user to use those tared modules to be sure a release
will always be working.