add virtualization tools
This commit adds the tools in order to be able to
bootstrap a virtual infrastructure based on libvirt.

The first tool is the collector which generate a virtual
description (in YAML), based on a real spinalstack deployment.

This file will be used by another tool which deploy the
infrastructure on libvirt.