download.sh: download & extract eDeploy metadatas
When running an environment without external eDeploy server, it's useful
to be able to prepare the eDeploy server with metadatas when
provisionning a new version.

Example: Spinal Stack is already running and we want to upgrade it to
the next release. When running provision.sh with -i option and this
patch, the eDeploy server will download the roles (install-server
included) and the metadatas, so it will be able to upgrade itself
out-of-the-box.

This is the last feature missing to be able to upgrade without manual
intervention. Before this commit, we had to update ourselves the
install-server outside the CI (the CI is using an upstream server so we
had no problem...).