extract-archive: untar only if archive.tar is here
Only extract archive.tar if the file in the machine. When possible, it's
more efficient to directly rsync the content of the tarball from the
remote host. This instead of doing a scp of the tarball then an untar in
the remote host.

(cherry picked from commit df20de4e4fce8b889e030caba5691ed7e861217e)