extract-archive: untar only if archive.tar is here
Only extract archive.tar if the file in the machine. When possible, it's
more efficient to directly rsync the content of the tarball from the
remote host. This instead of doing a scp of the tarball then an untar in
the remote host.

(cherry picked from commit 4c7e2452bd6eddb74d4f6a4dea04170724a63c89)