Removes old images from glance in inception.sh
If an image exists with the same name as the role, but has a different
hash, it will remain in place. Inception will then create a new image
with a different name, which leads to some issues.

This patch makes sure that the old image is removed before adding the
new one in glance