configure.sh: verify-servers.sh -x will never fail again
At this stage of the script, verify-servers.sh (serverspec) has already been run successfuly
on all the nodes for the steps. Since we are using 'set -e', we would not be here if
an error occured on the tests, so there is no reason to fail here.
With XML output, it happens serverspecs timeout to connect on the nodes,
so as a temporary background until we upgrade to latest version, do not fail on
this command to avoid deployment issues.