upgrade.sh: allow to continue the process after a stop
If there is an error in an Ansible playbook the process will immediately
stop to avoid any issue on the platform.

To continue the upgrade process without re-doing all the working steps,
let's use the same system as configure.sh does with Puppet, where each
successful step is written in a file, per host, per role.