javelin: don't run create_resources if all resources already exist
In the case we want to re-run Sanity process, but Javelin resources
already exist and are checked for the current version, we don't want to
run again the create_resources function, to avoid error during the run.

This patch re-factorize a bit variable and re-use the work to bring this
feature.